requirements.txt
argparse==1.4.0
astroid==1.5.3
backports.functools-lru-cache==1.4
colorama==0.3.9
enum34==1.1.6
funcsigs==1.0.2
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
mock==2.0.0
nose==1.3.7
ordereddict==1.1
pbr==3.1.1
pylint==1.7.2
scratchpy==0.1.0
singledispatch==3.4.0.3
six==1.10.0
wrapt==1.10.11
